metrics: move to subsystem
